extract model factory and stream parsing from agent/index.ts
